6597 change sub process id to code.

+++ dolphinscheduler-api/src/test/java/org/apache/dolphinscheduler/api/service/ExecutorServiceTest.java
@@ -279,8 +279,8 @@ public class ExecutorServiceTest {
 
     @Test
     public void testStartCheckByProcessDefinedCode() {
-        List<Integer> ids = new ArrayList<>();
-        ids.add(1);
+        List<Long> ids = new ArrayList<>();
+        ids.add(1L);
         Mockito.doNothing().when(processService).recurseFindSubProcessId(1, ids);
 
         List<ProcessDefinition> processDefinitionList = new ArrayList<>();
